Fixes/Improvements
Fixed the form submission for the contact page - on successful submission, users will be redirected to a new page called message.html.

Created footer styles and formatted a new footer design - LMK what you think.

Added a dynamic back-to-top button to every page - only appears when the user has scrolled down 100px.

Added smooth scrolling - Use the order.html in-page navigation to see how it works. Would love some feedback on that.

Changed the order.html in-page navbar to hold a cart link instead of back-to-top.